Linux amd64 patch
One thing I noticed that could be propagated to other platforms are all the getcallerpc-<ARCH>.S files under lib9 - gcc and clang have a builtin that do what should happen for that function, and it might be worth it to use that across the board for any host OS that uses gcc and/or clang. That was really the only hard part I ran into and took any major amount of research on my part.

John Osborne